requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.6
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tenacity==8.2.2
typing_extensions==4.7.1
uvicorn==0.23.1
